feat: Refactor user creation use case and update router import path
- Moved the create-user use case to a dedicated user directory for better organization.
- Updated the user router to reflect the new import path for the createUserUseCase function, enhancing code structure and maintainability.
